# api/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Viz API
  version: 0.1.0
  description: OpenAPI specification for the Viz API (chi-based)
servers:
  - url: http://{host}:{port}
    description: Local API
    variables:
      host:
        default: localhost
        description: Hostname or IP for the API server
      port:
        default: "7770"
        description: Port for the API server
  - url: /api
    description: Production API
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
...
  /auth/login:
    post:
      summary: Login with email and password
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: User email address
                password:
                  type: string
                  description: User password
              required: [email, password]
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/oauth:
    get:
      summary: Initiate OAuth flow
      operationId: initiateOAuth
      security: []
      parameters:
        - name: provider
          in: query
          required: true
          schema:
            type: string
            enum: [google, github]
          description: OAuth provider (google or github)
      responses:
        "307":
          description: Redirect to OAuth provider
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/oauth/{provider}:
    post:
      summary: Complete OAuth flow
      operationId: completeOAuth
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github]
          description: OAuth provider (google or github)
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: Authorization code from OAuth provider
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: State parameter for CSRF protection
      responses:
        "200":
          description: OAuth successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthUserData"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/session:
    get:
      summary: Get current session information
      operationId: getCurrentSession
      security:
        - CookieAuth: []
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      summary: Logout current session
      operationId: logout
      security:
        - CookieAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Unauthorized (no active session)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "API key for authentication. The key should be provided in the Authorization header as a Bearer token. This token is a hex-encoded string of random bytes, prefixed with 'IMAG_'. Bearer format is intentionally omitted to avoid JWT confusion."
      x-scopes:
        admin:read: Read admin resources
        admin:write: Write admin resources
        api_keys:create: Create API keys
        api_keys:list: List API keys
        api_keys:get: Get API key details
        api_keys:revoke: Revoke API keys
        api_keys:rotate: Rotate API keys
        api_keys:delete: Delete API keys
        auth:refresh: Refresh authentication tokens
        auth:revoke: Revoke authentication tokens
        collections:create: Create collections
        collections:read: Read collections
        collections:update: Update collections
        collections:delete: Delete collections
        collections:share: Share collections
        images:read: Read images
        images:upload: Upload images
        images:update: Update images
        images:delete: Delete images
        images:download: Download images
        jobs:read: Read jobs
        jobs:create: Create jobs
        jobs:update: Update jobs
        jobs:delete: Delete jobs
        users:create: Create users
        users:read: Read users
        users:update: Update users
        users:delete: Delete users
        user_settings:read: Read user settings
        user_settings:update: Update user settings
        downloads:create: Create downloads
        events:read: Read events
        "*": Grants all permissions (superuser)

    CookieAuth:
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          description: Response message
      required: [message]
...
    OAuthUserData:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User email
        name:
          type: string
          description: User name
        picture:
          type: string
          format: uri
          description: User profile picture URL
      required: [email, name, picture]
...
    User:
      x-entity: true
      type: object
      properties:
        uid: { type: string, description: User UID }
        first_name: { type: string, description: First name }
        last_name: { type: string, description: Last name }
        username: { type: string, description: Username }
        email: { type: string, description: Email }
        role:
          type: string
          enum: [user, admin, superadmin, guest]
          description: User role
        created_at:
          { type: string, format: date-time, description: Creation time }
        updated_at:
          { type: string, format: date-time, description: Update time }
      required:
        [
          uid,
          first_name,
          last_name,
          username,
          email,
          role,
          created_at,
          updated_at,
        ]
...
    Session:
      x-entity: true
      type: object
      properties:
        uid: { type: string, description: Session UID }
        token: { type: string, description: Session token }
        user_uid: { type: string, description: User UID }
        user:
          $ref: "#/components/schemas/User"
        client_id: { type: string, description: Client ID }
        client_name: { type: string, description: Client name }
        client_ip: { type: string, description: Client IP }
        last_active:
          { type: string, format: date-time, description: Last active time }
        expires_at:
          { type: string, format: date-time, description: Expiry time }
        timeout:
          { type: integer, format: int64, description: Timeout in seconds }
        user_agent: { type: string, description: User agent }
        ref_id: { type: string, description: Reference ID }
        login_ip: { type: string, description: Login IP }
        login_at: { type: string, format: date-time, description: Login time }
        status: { type: integer, description: Session status }
        created_at:
          { type: string, format: date-time, description: Creation time }
        updated_at:
          { type: string, format: date-time, description: Update time }
      required: [uid, token, user_uid, created_at, updated_at]
...